Se agrega nueva pantalla de cobros todo junto

+++ src/main/java/mx/com/tvch/pos/util/Constantes.java
@@ -76,19 +76,19 @@ public class Constantes {
     public static final int TIPO_BUSQUEDA_APELLIDO_MATERNO = 4;
     public static final int TIPO_BUSQUEDA_DOMICILIO = 5;
     
-    public static final int TIPO_ORDEN_INSTALACION = 1;
-    public static final int TIPO_ORDEN_SERVICIO = 2;
-    public static final int TIPO_ORDEN_CAMBIO_DOMICILIO = 4;
+    public static final long TIPO_ORDEN_INSTALACION = 1;
+    public static final long TIPO_ORDEN_SERVICIO = 2;
+    public static final long TIPO_ORDEN_CAMBIO_DOMICILIO = 4;
     
-    public static final int TIPO_ORDEN_SERVICIO_TV_ADICIONAL = 1;
-    public static final int TIPO_ORDEN_SERVICIO_CAMBIO_CABLE_ACOMETIDA = 2;
-    public static final int TIPO_ORDEN_SERVICIO_CAMBIO_CABLEADO = 3;
-    public static final int TIPO_ORDEN_SERVICIO_REUBICACION_MODEM = 4;
-    public static final int TIPO_ORDEN_SERVICIO_CAMBIO_UBICACION_TV = 5;
-    public static final int TIPO_ORDEN_SERVICIO_VISITA_TECNICA = 6;
-    public static final int TIPO_ORDEN_SERVICIO_CAMBIO_PLAN = 7;
-    public static final int TIPO_ORDEN_SERVICIO_RETIRO_EQUIPO_CANCELACION = 8;
-    public static final int TIPO_ORDEN_SERVICIO_RECONEXION_SERVICIO = 9;
+    public static final long TIPO_ORDEN_SERVICIO_TV_ADICIONAL = 1;
+    public static final long TIPO_ORDEN_SERVICIO_CAMBIO_CABLE_ACOMETIDA = 2;
+    public static final long TIPO_ORDEN_SERVICIO_CAMBIO_CABLEADO = 3;
+    public static final long TIPO_ORDEN_SERVICIO_REUBICACION_MODEM = 4;
+    public static final long TIPO_ORDEN_SERVICIO_CAMBIO_UBICACION_TV = 5;
+    public static final long TIPO_ORDEN_SERVICIO_VISITA_TECNICA = 6;
+    public static final long TIPO_ORDEN_SERVICIO_CAMBIO_PLAN = 7;
+    public static final long TIPO_ORDEN_SERVICIO_RETIRO_EQUIPO_CANCELACION = 8;
+    public static final long TIPO_ORDEN_SERVICIO_RECONEXION_SERVICIO = 9;
     
     public static final long TIPO_COBRO_ORDEN_INSTALACION = 1;
     public static final long TIPO_COBRO_SERVICIO = 2;
